+++ addons/l10n_si_pre_arrival_questionnaire/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import questionnaire_template
+from . import questionnaire_response
+from . import hotel_folio